+++ packages/openapi/tsup.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from 'tsup';
 
 export default defineConfig({
-  entry: ['./src/index.ts', './src/bin.ts'],
+  entry: ['./src/index.ts'],
   format: 'esm',
   dts: true,
   target: 'node18',