+++ packages/typescript/.eslintrc.cjs
@@ -1,0 +1,8 @@
+module.exports = {
+  extends: ['custom/library'],
+  ignorePatterns: ['*.test.ts'],
+  rules: {
+    // typescript is CommonJS
+    'import/no-named-as-default-member': 'off',
+  },
+};