+++ packages/ui/.eslintrc.cjs
@@ -1,6 +1,7 @@
 module.exports = {
   extends: ['custom/next'],
   rules: {
+    // for the import hacks
     '@typescript-eslint/consistent-type-imports': 'off',
   },
 };