+++ packages/ui/src/components/type-table.tsx
@@ -3,6 +3,7 @@
 import { InfoIcon } from 'lucide-react';
 import Link from 'next/link';
 import type { ReactNode } from 'react';
+import { cva } from 'class-variance-authority';
 import { cn } from '@/utils/cn';
 import {
   Popover,
@@ -16,7 +17,7 @@ export function Info({ children }: { children: ReactNode }): JSX.Element {
       <PopoverTrigger>
         <InfoIcon className="size-4" />
       </PopoverTrigger>
-      <PopoverContent className="text-sm">{children}</PopoverContent>
+      <PopoverContent className="prose text-sm">{children}</PopoverContent>
     </Popover>
   );
 }
@@ -25,7 +26,7 @@ interface ObjectType {
   /**
    * Additional description of the field
    */
-  description?: string;
+  description?: ReactNode;
   type: string;
   typeDescription?: string;
   /**
@@ -35,40 +36,42 @@ interface ObjectType {
   default?: string;
 }
 
+const th = cva('p-2 font-medium first:pl-0 last:pr-0');
+const td = cva('p-2 first:pl-0 last:pr-0');
+const field = cva('inline-flex flex-row items-center gap-1');
+const code = cva('rounded-md bg-secondary p-1 text-secondary-foreground', {
+  variants: {
+    color: { primary: 'bg-primary/10 text-primary' },
+  },
+});
+
 export function TypeTable({
   type,
 }: {
   type: Record<string, ObjectType>;
 }): JSX.Element {
-  const th = cn('p-2 font-medium first:pl-0 last:pr-0');
-  const td = cn('p-2 first:pl-0 last:pr-0');
-  const field = cn('inline-flex flex-row items-center gap-1');
-  const code = cn('rounded-md bg-secondary p-1 text-secondary-foreground');
-
   return (
     <div className="not-prose overflow-auto whitespace-nowrap">
       <table className="my-4 w-full text-left text-sm text-muted-foreground">
         <thead className="border-b">
           <tr>
-            <th className={cn(th, 'w-[45%]')}>Prop</th>
-            <th className={cn(th, 'w-[30%]')}>Type</th>
-            <th className={cn(th, 'w-[25%]')}>Default</th>
+            <th className={cn(th(), 'w-[45%]')}>Prop</th>
+            <th className={cn(th(), 'w-[30%]')}>Type</th>
+            <th className={cn(th(), 'w-[25%]')}>Default</th>
           </tr>
         </thead>
         <tbody className="border-collapse divide-y divide-border">
           {Object.entries(type).map(([key, value]) => (
             <tr key={key}>
-              <td className={td}>
-                <div className={field}>
-                  <code className={cn(code, 'bg-primary/10 text-primary')}>
-                    {key}
-                  </code>
+              <td className={td()}>
+                <div className={field()}>
+                  <code className={cn(code({ color: 'primary' }))}>{key}</code>
                   {value.description ? <Info>{value.description}</Info> : null}
                 </div>
               </td>
-              <td className={td}>
-                <div className={field}>
-                  <code className={code}>{value.type}</code>
+              <td className={td()}>
+                <div className={field()}>
+                  <code className={code()}>{value.type}</code>
                   {value.typeDescription ? (
                     <Info>
                       <pre className="overflow-auto bg-secondary text-secondary-foreground">
@@ -83,9 +86,9 @@ export function TypeTable({
                   ) : null}
                 </div>
               </td>
-              <td className={td}>
+              <td className={td()}>
                 {value.default ? (
-                  <code className={code}>{value.default}</code>
+                  <code className={code()}>{value.default}</code>
                 ) : (
                   <span>-</span>
                 )}