+++ packages/core/CHANGELOG.md
@@ -1,5 +1,24 @@
 # next-docs-zeta
 
+## 10.0.0
+
+### Major Changes
+
+- 321d1e1f: **Move Typescript integrations to `fumadocs-typescript`**
+
+  why: It is now a stable feature
+
+  migrate: Use `fumadocs-typescript` instead.
+
+  ```diff
+  - import { AutoTypeTable } from "fumadocs-ui/components/auto-type-table"
+  + import { AutoTypeTable } from "fumadocs-typescript/ui"
+  ```
+
+### Minor Changes
+
+- b5d16938: Support external link in `pages` property
+
 ## 9.1.0
 
 ### Minor Changes